Added types for lodash and lodash lib into the project. Refactore code for my navigation feature and created directory for future features.
