--- src/components/navigation/constants.ts
@@ -1,1 +1,0 @@
-
