+++ src/features/navigation/constants.ts
@@ -1,0 +1,7 @@
+import { ShoppingCart, Heart, User } from "lucide-react";
+
+export const menuListBattons = [
+	{ href: "#", icon: ShoppingCart, text: "Тотал" },
+	{ href: "#", icon: Heart, text: "Обране" },
+	{ href: "#", icon: User, text: "Профіль" },
+];
